Avoid inefficient `lastIndex`-operations in Navagator.
The `Navigator `and `_History` class performs some operations
by using `_lastRouteEntryWhereOrNull` which was written to
iterate from the start of the list to find the last entry
satisfying a predicate, which means iterating the entire
list every time.

The `popUntil` operation would repeatedly perform this linear
operation until finding an element which satisfies its
predicate.

Changed that to searching from the back, which should avoid
searching the entire list each time.
Also avoided doing the same work twice when `popUntil` calls `pop`.

The `popUntil` operation may still be quadratic since it starts
every search from the end.
(If there is some way to recognize that the history list hasn't
changed during the `pop` operation, the `popUntil` could start
searhching at the current position instead.)

Made `_History.addAll` not iterate its argument iterable more than once.

Added more `Iterable` methods to `_History` where the code was using
the default implementation, and the implementation could be more efficient
when based on a `List`. Not all possible optimizable members, just the
ones that are actually used.

Rewrote some `assert`s from
```
assert((){ if (test) { throw FlutterError(...); } return true; }());
```
to
```
assert(!test, throw FlutterError(...));
```
It's cleaner.

Typed `_allRouteOverlayEntries` as returning the list it is returning,
so that there is no need to call `.toList()` on it.